requirements/by_hitomi3_11requirements.txt
absl-py==2.3.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
anyascii==0.3.3
anyio==4.11.0
astunparse==1.6.3
attrs==25.4.0
audioread==3.1.0
av==16.0.1
babel==2.17.0
bangla==0.0.5
beautifulsoup4==4.14.2
blinker==1.9.0
blis==1.3.3
bnnumerizer==0.0.2
bnunicodenormalizer==0.1.7
bottle==0.13.4
Bottleneck==1.6.0
cachetools==6.2.2
catalogue==2.0.10
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloudpathlib==0.23.0
coacd==1.0.7
colorama==0.4.6
confection==0.1.5
contourpy==1.3.3
coqpit==0.0.17
cycler==0.12.1
cymem==2.0.13
Cython==3.2.2
dataclasses-json==0.6.7
dateparser==1.1.8
decorator==5.2.1
deepface==0.0.95
Deprecated==1.3.1
dill==0.4.0
docopt==0.6.2
docstring_parser==0.17.0
docutils==0.22.3
etils==1.13.0
fast_simplification==0.1.12
ffpyplayer==4.5.3
filelock==3.20.0
filetype==1.2.0
fire==0.7.1
Flask==3.1.2
flask-cors==6.0.1
flatbuffers==25.9.23
fonttools==4.61.0
freetype-py==2.5.1
frozenlist==1.8.0
fsspec==2025.10.0
g2pkk==0.1.2
gast==0.6.0
gdown==5.2.0
genesis-world==0.3.7
glfw==2.10.0
google-ai-generativelanguage==0.9.0
google-api-core==2.28.1
google-auth==2.43.0
google-cloud-aiplatform==1.128.0
google-cloud-bigquery==3.38.0
google-cloud-core==2.5.0
google-cloud-resource-manager==1.15.0
google-cloud-storage==3.6.0
google-crc32c==1.7.1
google-genai==1.52.0
google-pasta==0.2.0
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
greenlet==3.2.4
grpc-google-iam-v1==0.14.3
grpcio==1.76.0
grpcio-status==1.76.0
gruut==2.2.3
gruut-ipa==0.13.0
gruut_lang_de==2.0.1
gruut_lang_en==2.0.1
gruut_lang_es==2.0.1
gruut_lang_fr==2.0.2
gs-madrona==0.0.5
gstaichi==4.0.0
gunicorn==23.0.0
h11==0.16.0
h5py==3.15.1
hangul-romanize==0.1.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.36.0
idna==3.11
ImageIO==2.37.2
imageio-ffmpeg==0.6.0
importlib_resources==6.5.2
inflect==7.5.0
itsdangerous==2.2.0
jamo==0.4.1
jieba==0.42.1
Jinja2==3.1.6
joblib==1.5.2
jsonlines==1.2.0
jsonpatch==1.33
jsonpointer==3.0.0
keras==3.12.0
Kivy==2.3.1
Kivy-Garden==0.1.5
kiwisolver==1.4.9
langchain==1.1.0
langchain-classic==1.0.0
langchain-community==0.4.1
langchain-core==1.1.0
langchain-google-genai==3.2.0
langchain-google-vertexai==3.1.0
langchain-text-splitters==1.0.0
langgraph==1.0.4
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.5
langgraph-sdk==0.2.10
langsmith==0.4.49
lazy_loader==0.4
libclang==18.1.1
libigl==2.6.1
librosa==0.11.0
llvmlite==0.45.1
lz4==4.4.5
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.1
matplotlib==3.10.7
mdurl==0.1.2
ml_dtypes==0.5.4
more-itertools==10.8.0
moviepy==2.2.1
msgpack==1.1.2
mtcnn==1.0.0
mujoco==3.3.7
multidict==6.7.0
murmurhash==1.0.15
mypy_extensions==1.1.0
namex==0.1.0
networkx==2.8.8
nltk==3.9.2
num2words==0.5.14
numba==0.62.1
numexpr==2.14.1
numpy==1.26.4
opencv-python==4.12.0.88
OpenEXR==3.4.4
opt_einsum==3.4.0
optree==0.18.0
orjson==3.11.4
ormsgpack==1.12.0
packaging==25.0
pandas==1.5.3
pillow==11.3.0
pip-autoremove==0.10.0
platformdirs==4.5.0
plotly==5.24.1
pooch==1.8.2
preshed==3.0.12
proglog==0.1.12
propcache==0.4.1
proto-plus==1.26.1
protobuf==6.33.1
proxy_tools==0.1.0
psutil==7.1.3
py-cpuinfo==9.0.0
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycairo==1.29.0
pycollada==0.9.2
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
PyGEL3D==0.6.1
pyglet==2.1.11
pygltflib==1.16.0
Pygments==2.19.2
PyGObject==3.50.0
PyGObject-stubs==2.14.0
pymeshlab==2025.7
pynndescent==0.5.13
PyOpenGL==3.1.10
pyparsing==3.2.5
pypinyin==0.55.0
PyQt6==6.10.0
PyQt6-Qt6==6.10.0
PyQt6_sip==13.10.2
pysbd==0.3.4
PySocks==1.7.1
pysplashsurf==0.12.0.2
python-crfsuite==0.9.11
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
pyvista==0.46.4
pywebview==6.1
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
retina-face==0.0.17
rich==14.2.0
rsa==4.9.1
rtree==1.4.1
safetensors==0.7.0
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.16.3
scooby==0.11.0
shapely==2.1.2
six==1.17.0
smart_open==7.5.0
sniffio==1.3.1
sounddevice==0.5.3
soundfile==0.13.1
soupsieve==2.8
soxr==1.0.0
spacy==3.8.11
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.44
srsly==2.5.2
SudachiDict-core==20251022
SudachiPy==0.6.10
tenacity==9.1.2
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.2.0
tetgen==0.7.0
thinc==8.3.10
threadpoolctl==3.6.0
tifffile==2025.10.16
tokenizers==0.20.0
tqdm==4.67.1
trainer==0.0.36
transformers==4.46.2
trimesh==4.10.0
typeguard==4.4.4
typer-slim==0.20.0
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
umap-learn==0.5.9.post2
Unidecode==1.4.0
urllib3==2.5.0
validators==0.35.0
vtk==9.5.2
wasabi==1.1.3
weasel==0.4.3
websockets==15.0.1
Werkzeug==3.1.3
wrapt==2.0.1
xxhash==3.6.0
yarl==1.22.0
z3-solver==4.15.4.0
zipp==3.23.0
zstandard==0.25.0
